wip: see if this passes

+++ core/geotime/src/main/java/org/locationtech/geowave/core/geotime/store/statistics/TimeRangeDataStatistics.java
@@ -76,8 +76,8 @@ abstract public class TimeRangeDataStatistics<T> extends
 	@Override
 	public byte[] toBinary() {
 		final ByteBuffer buffer = super.binaryBuffer(16);
-		buffer.putDouble(min);
-		buffer.putDouble(max);
+		buffer.putLong(min);
+		buffer.putLong(max);
 		return buffer.array();
 	}
